Desktop/โปรเจคสหกิจ/project1/requirements.txt
sentence-transformers
faiss-cpu
openai
langchain-text-splitters
python-dotenv
